passport-shop-v2/requirements.txt
pillow
numpy
mediapipe
scipy